move business logic into ticket service, refactor code, add more test
