+++ src/app.module.ts
@@ -4,10 +4,11 @@ import { TicketsController } from './tickets/tickets.controller';
 import { ReportsController } from './reports/reports.controller';
 import { HealthcheckController } from './healthcheck/healthcheck.controller';
 import { ReportsService } from './reports/reports.service';
+import { TicketsService } from './tickets/tickets.service';
 
 @Module({
   imports: [DbModule],
   controllers: [TicketsController, ReportsController, HealthcheckController],
-  providers: [ReportsService],
+  providers: [ReportsService, TicketsService],
 })
 export class AppModule {}
