+++ src/tickets/tickets.controller.spec.ts
@@ -9,6 +9,7 @@ import {
 import { User, UserRole } from '../../db/models/User';
 import { DbModule } from '../db.module';
 import { TicketsController } from './tickets.controller';
+import { TicketsService } from './tickets.service';
 
 describe('TicketsController', () => {
   let controller: TicketsController;
@@ -16,6 +17,7 @@ describe('TicketsController', () => {
   beforeEach(async () => {
     const module: TestingModule = await Test.createTestingModule({
       controllers: [TicketsController],
+      providers: [TicketsService],
       imports: [DbModule],
     }).compile();
 
@@ -27,125 +29,5 @@ describe('TicketsController', () => {
 
     const res = await controller.findAll();
     console.log(res);
-  });
-
-  describe('create', () => {
-    describe('managementReport', () => {
-      it('creates managementReport ticket', async () => {
-        const company = await Company.create({ name: 'test' });
-        const user = await User.create({
-          name: 'Test User',
-          role: UserRole.accountant,
-          companyId: company.id,
-        });
-
-        const ticket = await controller.create({
-          companyId: company.id,
-          type: TicketType.managementReport,
-        });
-
-        expect(ticket.category).toBe(TicketCategory.accounting);
-        expect(ticket.assigneeId).toBe(user.id);
-        expect(ticket.status).toBe(TicketStatus.open);
-      });
-
-      it('if there are multiple accountants, assign the last one', async () => {
-        const company = await Company.create({ name: 'test' });
-        await User.create({
-          name: 'Test User',
-          role: UserRole.accountant,
-          companyId: company.id,
-        });
-        const user2 = await User.create({
-          name: 'Test User',
-          role: UserRole.accountant,
-          companyId: company.id,
-        });
-
-        const ticket = await controller.create({
-          companyId: company.id,
-          type: TicketType.managementReport,
-        });
-
-        expect(ticket.category).toBe(TicketCategory.accounting);
-        expect(ticket.assigneeId).toBe(user2.id);
-        expect(ticket.status).toBe(TicketStatus.open);
-      });
-
-      it('if there is no accountant, throw', async () => {
-        const company = await Company.create({ name: 'test' });
-
-        await expect(
-          controller.create({
-            companyId: company.id,
-            type: TicketType.managementReport,
-          }),
-        ).rejects.toEqual(
-          new ConflictException(
-            `Cannot find user with role accountant to create a ticket`,
-          ),
-        );
-      });
-    });
-
-    describe('registrationAddressChange', () => {
-      it('creates registrationAddressChange ticket', async () => {
-        const company = await Company.create({ name: 'test' });
-        const user = await User.create({
-          name: 'Test User',
-          role: UserRole.corporateSecretary,
-          companyId: company.id,
-        });
-
-        const ticket = await controller.create({
-          companyId: company.id,
-          type: TicketType.registrationAddressChange,
-        });
-
-        expect(ticket.category).toBe(TicketCategory.corporate);
-        expect(ticket.assigneeId).toBe(user.id);
-        expect(ticket.status).toBe(TicketStatus.open);
-      });
-
-      it('if there are multiple secretaries, throw', async () => {
-        const company = await Company.create({ name: 'test' });
-        await User.create({
-          name: 'Test User',
-          role: UserRole.corporateSecretary,
-          companyId: company.id,
-        });
-        await User.create({
-          name: 'Test User',
-          role: UserRole.corporateSecretary,
-          companyId: company.id,
-        });
-
-        await expect(
-          controller.create({
-            companyId: company.id,
-            type: TicketType.registrationAddressChange,
-          }),
-        ).rejects.toEqual(
-          new ConflictException(
-            `Multiple users with role corporateSecretary. Cannot create a ticket`,
-          ),
-        );
-      });
-
-      it('if there is no secretary, throw', async () => {
-        const company = await Company.create({ name: 'test' });
-
-        await expect(
-          controller.create({
-            companyId: company.id,
-            type: TicketType.registrationAddressChange,
-          }),
-        ).rejects.toEqual(
-          new ConflictException(
-            `Cannot find user with role corporateSecretary to create a ticket`,
-          ),
-        );
-      });
-    });
   });
 });
