+++ src/tickets/tickets.service.spec.ts
@@ -1,0 +1,239 @@
+import { ConflictException } from '@nestjs/common';
+import { Test, TestingModule } from '@nestjs/testing';
+import { Company } from '../../db/models/Company';
+import {
+  Ticket,
+  TicketCategory,
+  TicketStatus,
+  TicketType,
+} from '../../db/models/Ticket';
+import { User, UserRole } from '../../db/models/User';
+import { DbModule } from '../db.module';
+import { TicketsService } from './tickets.service';
+
+describe('TicketsService', () => {
+  let service: TicketsService;
+
+  beforeEach(async () => {
+    const module: TestingModule = await Test.createTestingModule({
+      providers: [TicketsService],
+      imports: [DbModule],
+    }).compile();
+
+    service = module.get<TicketsService>(TicketsService);
+  });
+
+  it('should be defined', async () => {
+    expect(service).toBeDefined();
+  });
+
+  describe('create', () => {
+    describe('managementReport', () => {
+      it('creates managementReport ticket', async () => {
+        const company = await Company.create({ name: 'test' });
+        const user = await User.create({
+          name: 'Test User',
+          role: UserRole.accountant,
+          companyId: company.id,
+        });
+
+        const ticket = await service.createTicket({
+          companyId: company.id,
+          type: TicketType.managementReport,
+        });
+
+        expect(ticket.category).toBe(TicketCategory.accounting);
+        expect(ticket.assigneeId).toBe(user.id);
+        expect(ticket.status).toBe(TicketStatus.open);
+      });
+
+      it('if there are multiple accountants, assign the last one', async () => {
+        const company = await Company.create({ name: 'test' });
+        await User.create({
+          name: 'Test User',
+          role: UserRole.accountant,
+          companyId: company.id,
+        });
+        const user2 = await User.create({
+          name: 'Test User',
+          role: UserRole.accountant,
+          companyId: company.id,
+        });
+
+        const ticket = await service.createTicket({
+          companyId: company.id,
+          type: TicketType.managementReport,
+        });
+
+        expect(ticket.category).toBe(TicketCategory.accounting);
+        expect(ticket.assigneeId).toBe(user2.id);
+        expect(ticket.status).toBe(TicketStatus.open);
+      });
+
+      it('if there is no accountant, throw', async () => {
+        const company = await Company.create({ name: 'test' });
+
+        await expect(
+          service.createTicket({
+            companyId: company.id,
+            type: TicketType.managementReport,
+          }),
+        ).rejects.toEqual(
+          new ConflictException(
+            `Cannot find user with role accountant to create a ticket`,
+          ),
+        );
+      });
+    });
+
+    describe('registrationAddressChange', () => {
+      it('creates registrationAddressChange ticket', async () => {
+        const company = await Company.create({ name: 'test' });
+        const user = await User.create({
+          name: 'Test User',
+          role: UserRole.corporateSecretary,
+          companyId: company.id,
+        });
+
+        const ticket = await service.createTicket({
+          companyId: company.id,
+          type: TicketType.registrationAddressChange,
+        });
+
+        expect(ticket.category).toBe(TicketCategory.corporate);
+        expect(ticket.assigneeId).toBe(user.id);
+        expect(ticket.status).toBe(TicketStatus.open);
+      });
+
+      it('if there are multiple secretaries, throw', async () => {
+        const company = await Company.create({ name: 'test' });
+        await User.create({
+          name: 'Test User',
+          role: UserRole.corporateSecretary,
+          companyId: company.id,
+        });
+        await User.create({
+          name: 'Test User',
+          role: UserRole.corporateSecretary,
+          companyId: company.id,
+        });
+
+        await expect(
+          service.createTicket({
+            companyId: company.id,
+            type: TicketType.registrationAddressChange,
+          }),
+        ).rejects.toEqual(
+          new ConflictException(
+            `Multiple users with role corporateSecretary. Cannot create a ticket`,
+          ),
+        );
+      });
+
+      it('if there is no secretary, assign it to director', async () => {
+        const company = await Company.create({ name: 'test' });
+        const user = await User.create({
+          name: 'Test User',
+          role: UserRole.director,
+          companyId: company.id,
+        });
+
+        const ticket = await service.createTicket({
+          companyId: company.id,
+          type: TicketType.registrationAddressChange,
+        });
+
+        expect(ticket.category).toBe(TicketCategory.corporate);
+        expect(ticket.assigneeId).toBe(user.id);
+        expect(ticket.status).toBe(TicketStatus.open);
+      });
+
+      it('if company has registrationAddressChange ticket, cannot create another one', async () => {
+        const company = await Company.create({ name: 'test' });
+        const user = await User.create({
+          name: 'Test User',
+          role: UserRole.corporateSecretary,
+          companyId: company.id,
+        });
+        await Ticket.create({
+          companyId: company.id,
+          assigneeId: user.id,
+          category: TicketCategory.corporate,
+          type: TicketType.registrationAddressChange,
+          status: TicketStatus.open,
+        });
+
+        await expect(
+          service.createTicket({
+            companyId: company.id,
+            type: TicketType.registrationAddressChange,
+          }),
+        ).rejects.toEqual(
+          new ConflictException(
+            `Cannot create a ticket with multiple type of ${TicketType.registrationAddressChange} for company ${company.id}`,
+          ),
+        );
+      });
+
+      it('if multiple users with director role, cannot create a ticket', async () => {
+        const company = await Company.create({ name: 'test' });
+        await User.create({
+          name: 'Test User',
+          role: UserRole.director,
+          companyId: company.id,
+        });
+        await User.create({
+          name: 'Test User2',
+          role: UserRole.director,
+          companyId: company.id,
+        });
+
+        await expect(
+          service.createTicket({
+            companyId: company.id,
+            type: TicketType.registrationAddressChange,
+          }),
+        ).rejects.toEqual(
+          new ConflictException(
+            `Multiple users with role ${UserRole.director}. Cannot create a ticket`,
+          ),
+        );
+      });
+    });
+
+    describe('strikeOff', () => {
+      it('resolve all tickets if ticket type is strikeOff', async () => {
+        const company = await Company.create({ name: 'test' });
+        const user = await User.create({
+          name: 'Test User',
+          role: UserRole.corporateSecretary,
+          companyId: company.id,
+        });
+        const directorUser = await User.create({
+          name: 'Test Director User',
+          role: UserRole.director,
+          companyId: company.id,
+        });
+        const ticket = await Ticket.create({
+          companyId: company.id,
+          assigneeId: user.id,
+          category: TicketCategory.corporate,
+          type: TicketType.registrationAddressChange,
+          status: TicketStatus.open,
+        });
+
+        const ticketDto = await service.createTicket({
+          companyId: company.id,
+          type: TicketType.strikeOff,
+        });
+
+        const ResolvedTicket = await Ticket.findOne({
+          where: { id: ticket.id },
+        });
+
+        expect(ticketDto.assigneeId).toBe(directorUser.id);
+        expect(ResolvedTicket?.status).toBe(TicketStatus.resolved);
+      });
+    });
+  });
+});
